requirements.txt
python-dotenv==1.0.0
python-dateutil==2.8.2
requests